requirements.txt
numpy
matplotlib
sqlalchemy
pandas
python-dotenv
seaborn
psycopg2
